+++ packages/core/src/util/initializedTooLateHeuristic.js
@@ -1,7 +1,14 @@
 'use strict';
 
+let logger;
+logger = require('../logger').getLogger('util/initializedTooLateHeuristic', newLogger => {
+  logger = newLogger;
+});
+
 let firstCall = true;
-const patterns = [
+let hasBeenInitializedTooLate = false;
+
+let patterns = [
   /\/@elastic\/elasticsearch\/index.js/,
   /\/@google-cloud\/pubsub\/build\/src\/publisher\/index.js/,
   /\/@google-cloud\/pubsub\/build\/src\/subscriber.js/,
@@ -9,7 +16,6 @@ const patterns = [
   /\/@hapi\/call\/lib\//,
   /\/amqplib\/lib\//,
   /\/aws-sdk\/lib\/aws.js/,
-  /\/bluebird\/js\/release\//,
   // deliberately not including bunyan because we depend on bunyan ourselves
   /\/elasticsearch\/src\/elasticsearch.js/,
   /\/express\/index.js/,
@@ -34,22 +40,42 @@ const patterns = [
   /\/pg\/lib\//,
   /\/pino\/lib\//,
   /\/redis\/index.js/,
-  /\/request\/index.js/,
-  /\/superagent\/lib\/node\/index.js/,
-  /\/winston\/lib\/winston.js/
+  /\/superagent\/lib\/node\/index.js/
 ];
 
-let hasBeenInitializedTooLate = false;
+const extraPatterns = [
+  // The following patterns are excluded when @contrast/agent (which also comes with the requirement to be loaded before
+  // everything else) is already loaded before @instana/core. This is to avoid false positive warnings about
+  // @instana/core being initialized too late.
+  /\/bluebird\/js\/release\//,
+  /\/request\/index.js/,
+  /\/winston\/lib\/winston.js/
+];
 
 module.exports = exports = function hasThePackageBeenInitializedTooLate() {
   if (firstCall) {
     const loadedModules = Object.keys(require.cache);
-    // eslint-disable-next-line no-restricted-syntax
-    outer: for (let i = 0; i < loadedModules.length; i++) {
+
+    // Run a pre-check to avoid the warning in the presence of modules that can legitimately be loaded before.
+    const addExtraPatterns =
+      Object.keys(require.cache).filter(moduleId => /\/@contrast\/agent\//.test(moduleId)).length === 0;
+    if (addExtraPatterns) {
+      patterns = patterns.concat(extraPatterns);
+    } else {
+      logger.debug(
+        // eslint-disable-next-line max-len
+        'Found @contrast/agent in the modules that have already been loaded. @instana/core will therefore exclude bluebird, request and winston from the check for modules that have been loaded before @instana/core.'
+      );
+    }
+
+    for (let i = 0; i < loadedModules.length; i++) {
       for (let j = 0; j < patterns.length; j++) {
         if (patterns[j].test(loadedModules[i])) {
           hasBeenInitializedTooLate = true;
-          break outer;
+          logger.debug(
+            // eslint-disable-next-line max-len
+            `Found a module that has been loaded before @instana/core but should have been loaded afterwards: ${loadedModules[i]}.`
+          );
         }
       }
     }